+++ .storybook/stories/consent.stories.ts
@@ -5,12 +5,12 @@ import { html } from "lit";
 import "../../src/hotosm-ui";
 
 const meta: Meta = {
-  title: "Tracking",
-  component: "hot-tracking",
+  title: "Consent",
+  component: "hot-consent",
 };
 export default meta;
 
-export const TrackingMatomo: StoryObj = {
+export const Consent: StoryObj = {
   args: {
     siteId: "1",
     domain: "localhost",
@@ -22,31 +22,25 @@ export const TrackingMatomo: StoryObj = {
         type: "select",
       },
     },
-    domain: {
-      options: ["localhost", "anotherdomain"],
-      control: {
-        type: "select",
-      },
-    },
   },
   parameters: {
     showAgreeToast: () => {
       const agree = document.getElementById("agree-toast");
       if (agree) {
-        agree.toast();
+        // agree.toast();
       }
     },
     showDisagreeToast: () => {
       const disagree = document.getElementById("disagree-toast");
       if (disagree) {
-        disagree.toast();
+        // disagree.toast();
       }
     },
     addKeyLocalStorage: (siteId: number) => {
-      localStorage.setItem(`${siteId}-matomo-agree`, 'true');
+      localStorage.setItem(`${siteId}-consent-agree`, 'true');
     },
     removeKeyLocalStorage: (siteId: number) => {
-      localStorage.removeItem(`${siteId}-matomo-agree`);
+      localStorage.removeItem(`${siteId}-consent-agree`);
     },
   },
   render: (args, { parameters }) => {
@@ -58,17 +52,30 @@ export const TrackingMatomo: StoryObj = {
         parameters.addKeyLocalStorage(args.siteId)
       }}>Disable Banner</sl-button>
       <br /><br />
-      <hot-matomo-tracking
-        id="matomo-banner"
-        site-id="${args.siteId}"
-        domain="${args.domain}"
+
+      <hot-consent
+
+      site-id=${args.siteId}
+        agree-label="Yes, I accept"
+        not-agree-label="I DO NOT accept"
+        title=${"What info we collect about you?"}
         @agree=${() => {
           parameters.showAgreeToast()
         }}
         @disagree=${() => {
           parameters.showDisagreeToast()
         }}
-      ></hot-matomo-tracking>
+      >
+        We use cookies and similar technologies to recognize and analyze your
+        visits, and measure traffic usage and activity. You can learn about how
+        we use the data about your visit or information you provide reading our
+        <a
+          href="https://www.hotosm.org/privacy"
+          target="_blank"
+          rel="noopener noreferrer"
+          >privacy policy</a
+        >. By clicking "I Agree", you consent to the use of cookies.
+      </hot-consent>
 
       <sl-alert id="agree-toast" variant="success" duration="3000" closable>
         <sl-icon slot="icon" name="check2-circle"></sl-icon>
