+++ src/components/consent/consent.styles.ts
@@ -1,19 +1,20 @@
-import{css as o}from"lit";var a=o`
-  .matomoTracking {
+import { css } from 'lit';
+
+export default css`
+  .consent {
     text-align: center;
   }
 
-  .matomoTracking--header {
+  .consent--header {
     font-weight: var(--hot-font-weight-semibold);
     font-size: var(--hot-font-size-large);
-    text-align: center;
   }
 
-  .matomoTracking--link {
+  .consent--link {
     color: var(--hot-color-primary-700);
   }
 
-  .matomoTracking--message {
+  .consent--message {
     margin-bottom: var(--hot-spacing-large);
   }
-`;export{a};
+`
