+++ src/components/consent/consent.ts
@@ -1,0 +1,11 @@
+import Consent from './consent.component.js';
+
+export * from './consent.component.js';
+export default Consent;
+
+declare global {
+  interface HTMLElementTagNameMap {
+    'hot-consent': Consent;
+  }
+}
+
