--- src/components/matomoTracking/matomoTracking.styles.ts
@@ -1,21 +1,0 @@
-import { css } from 'lit';
-
-export default css`
-  .matomoTracking {
-    text-align: center;
-  }
-
-  .matomoTracking--header {
-    font-weight: var(--hot-font-weight-semibold);
-    font-size: var(--hot-font-size-large);
-    text-align: center;
-  }
-
-  .matomoTracking--link {
-    color: var(--hot-color-primary-700);
-  }
-
-  .matomoTracking--message {
-    margin-bottom: var(--hot-spacing-large);
-  }
-`
