--- src/components/matomoTracking/matomoTracking.ts
@@ -1,11 +1,0 @@
-import MatomoTracking from './matomoTracking.component.js';
-
-export * from './matomoTracking.component.js';
-export default MatomoTracking;
-
-declare global {
-  interface HTMLElementTagNameMap {
-    'hot-matomo-tracking': MatomoTracking;
-  }
-}
-
