--- src/components/tracking/tracking.styles.ts
@@ -1,20 +1,0 @@
-import { css } from 'lit';
-
-export default css`
-  .tracking {
-    text-align: center;
-  }
-
-  .tracking--header {
-    font-weight: var(--hot-font-weight-semibold);
-    font-size: var(--hot-font-size-large);
-  }
-
-  .tracking--link {
-    color: var(--hot-color-primary-700);
-  }
-
-  .tracking--message {
-    margin-bottom: var(--hot-spacing-large);
-  }
-`
