+++ src/react/matomoTracking/index.ts
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { createComponent } from '@lit/react';
-import Component from '../../components/matomoTracking/matomoTracking.component.js';
+import Component from '../../components/tracking/tracking.component.js';
 
 const tagName = 'hot-matomo-tracking';
